word net test

+++ hammer-colombo/src/main/java/org/hammer/colombo/App.java
@@ -42,7 +42,7 @@ public class App {
 	 * @param thQuery the limit used to prunning the set of query (using cos(theta) function)
 	 * @throws Exception
 	 */
-	public static void Run(String fileQuery, String fileSystem, String searchMode, String queryMode, float thKrm,  float thRm, float thSim, int maxSim, String datasetTable, String indexTable, float thQuery) throws Exception {
+	public static void Run(String fileQuery, String fileSystem, String searchMode, String queryMode, float thKrm,  float thRm, float thSim, int maxSim, String datasetTable, String indexTable, int thQuery) throws Exception {
 		System.out.println("!!! Hammer Project !!!");
 		System.out.println("!!! Colombo Module start.....");
 		
@@ -161,7 +161,7 @@ public class App {
 		if (pArgs == null || pArgs.length < 11) {
 			throw new Exception("Parameter: <path_to_query> <file-system: local|hdfs> <search mode: search|download> <query mode: keywords|labels> <thKrm: 0.5|0.01..>  <thRm: 0.5|0.01..> <thSim: 0.5|0.01..> <maxSim: 1|2|3...> <dataset-table> <index-table> <cosTh:0.9991|0.9992>");
 		}
-		Run(pArgs[0], pArgs[1], pArgs[2], pArgs[3], Float.parseFloat(pArgs[4]), Float.parseFloat(pArgs[5]), Float.parseFloat(pArgs[6]), Integer.parseInt(pArgs[7]), pArgs[8], pArgs[9], Float.parseFloat(pArgs[10]));
+		Run(pArgs[0], pArgs[1], pArgs[2], pArgs[3], Float.parseFloat(pArgs[4]), Float.parseFloat(pArgs[5]), Float.parseFloat(pArgs[6]), Integer.parseInt(pArgs[7]), pArgs[8], pArgs[9], Integer.parseInt(pArgs[10]));
 	}
 
 	public static String ReadFileFromHdfs(Configuration conf) {
